Update README.md and RAL_Compiler.drawio with enhanced diagrams
Update the Functional and Class diagrams in the README.md to improve readability. Special characters in the filenames of these diagrams have been removed, again for readability.
Also, introduced a new section, "CPU Architecture," with an accompanying diagram, to provide additional documentation for the underlying CPU setup in the project.
No changes in functionalities were made, these changes are purely for documentation purposes.